Re-combine the elasticsearch module into the es5-persistence module.
Now that the es2-persistence module has been removed there isn't really a need to have a higher level elastic search module for shared code.  I left the more abstract package structure in place for now because I think that it still makes sense.
