+++ es5-persistence/src/main/java/com/netflix/conductor/elasticsearch/EmbeddedElasticSearch.java
