+++ es5-persistence/src/main/java/com/netflix/conductor/elasticsearch/query/parser/AbstractNode.java
