+++ es5-persistence/src/test/java/com/netflix/conductor/elasticsearch/query/parser/TestName.java
@@ -18,9 +18,6 @@
  */
 package com.netflix.conductor.elasticsearch.query.parser;
 
-import com.netflix.conductor.elasticsearch.query.parser.BooleanOp;
-import com.netflix.conductor.elasticsearch.query.parser.ParserException;
-
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -30,26 +27,14 @@ import static org.junit.Assert.assertNotNull;
  * @author Viren
  *
  */
-public class TestBooleanOp extends AbstractParserTest {
+public class TestName extends AbstractParserTest {
 
 	@Test
-	public void test() throws Exception {
-		String[] tests = new String[]{"AND", "OR"};
-		for(String test : tests){
-			BooleanOp name = new BooleanOp(getInputStream(test));
-			String nameVal = name.getOperator();
-			assertNotNull(nameVal);
-			assertEquals(test, nameVal);
-		}
-	}
-	
-	@Test(expected=ParserException.class)
-	public void testInvalid() throws Exception {
-		String test = "<";
-		BooleanOp name = new BooleanOp(getInputStream(test));
-		String nameVal = name.getOperator();
+	public void test() throws Exception{
+		String test =  "metadata.en_US.lang		";
+		Name name = new Name(getInputStream(test));
+		String nameVal = name.getName();
 		assertNotNull(nameVal);
-		assertEquals(test, nameVal);
-	
+		assertEquals(test.trim(), nameVal);
 	}
 }
